+++ observability/grafana-dashboards/bitmagnet.json
@@ -209,7 +209,7 @@
           "refId": "A"
         }
       ],
-      "title": "Torrents persisted per minute",
+      "title": "DHT crawler torrents persisted per minute",
       "type": "timeseries"
     },
     {
@@ -1313,6 +1313,6 @@
   "timezone": "",
   "title": "bitmagnet",
   "uid": "c0abadd0-6633-40c4-a2fc-01408a8dfc4e",
-  "version": 9,
+  "version": 1,
   "weekStart": ""
 }
